Extract src/debug.ts and src/dom.ts
The debug console — ring buffer, telemetry counters, badge, panel and the
snapshot flush — moves out whole. Unlike the three *view.ts modules it owns its
DOM region and its own state, so it exports render functions rather than markup;
main.ts keeps the button listeners and the 4s flush interval. The moved block is
byte-identical to HEAD:src/main.ts apart from `export` keywords and the
appVersion setter.

This is the slice that forced `$` into its own module. A DOM-owning module needs
document.getElementById and nothing else from main.ts, and importing main.ts is
what the dependency direction forbids — so ./dom exists to hold one line. Every
dialog still to be extracted will need it.

Three more things moved to where they belong, all of them read by the snapshot:

- `termEngine` -> ./state, a persisted preference exactly like sortMode and
  wtGroup. The Engine *type* goes to ./types with the other model types; the
  label table, availability probe and setEngine stay in main.ts, which is what
  offers them. Note setEngine still persists after calling setTermEngine — the
  imported binding is live, so it reads back the value just set.
- the dirty-by-folder cache (+ isDirty / folderDirty) -> ./state. It is keyed by
  folder rather than session on purpose: Sess.git only stays fresh for the pane
  on stage, so the sidebar dot, the external inspector and the snapshot all need
  this instead.

Verified in the running page rather than a fresh import, which is a better check
than the earlier slices got: clicking the real 🐞 button opens and closes the
panel, the badge already read 9 issues from startup logging, the telemetry
rx/routed/dropped line renders, and "app started" is in the ring — all through
the extracted module as main.ts wired it. dbgPath stays empty because
write_debug_file has no backend in a browser, which flushDebug catches.
